third_party/blink/web_tests/external/wpt/content-security-policy/reporting-api/reporting-api-sends-reports-on-violation.https.sub-expected.txt

This is a testharness.js-based test.
[FAIL] Report is observable to ReportingObserver
  assert_equals: expected (string) "https://web-platform.test:8444/content-security-policy/reporting-api/reporting-api-sends-reports-on-violation.https.sub.html" but got (object) null
Harness: the test ran to completion.
